break out helper functions from emissions line graph and remove filtering of lines in tooltip
